COM-1000 change how to getusers for coach table
